# spec/edited/openapi.yaml
openapi: 3.0.2
info:
  title: Wellbore DDMS OSDU
  description: build local
  version: '0.2'
paths:
...
  /ddms/v2/dipsets:
    post:
      tags:
        - Dipset
      summary: 'Create or update the DipSets using wks:dipSet:1.0.0 schema'
      operationId: post_dipset
      parameters:
        - description: identifier of the data partition to query
          required: false
          schema:
            title: data partition id
            minLength: 1
            type: string
            description: identifier of the data partition to query
          name: data-partition-id
          in: header
      requestBody:
        content:
          application/json:
            schema:
              title: Dipsets
              type: array
              items:
                $ref: '#/components/schemas/dipset'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateUpdateRecordsResponse'
        '400':
          description: Missing mandatory parameter or unknown parameter
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - OpenDESBearerToken: []
  '/ddms/v2/dipsets/{dipsetid}/versions/{version}':
    get:
      tags:
        - Dipset
      summary: 'Get the given version of DipSet using wks:dipset:1.0.0 schema'
      description: '"Get the DipSet object using its **id**.'
      operationId: get_dipset_version
      parameters:
        - required: true
          schema:
            title: Dipsetid
            type: string
          name: dipsetid
          in: path
        - required: true
          schema:
            title: Version
            type: integer
          name: version
          in: path
        - description: identifier of the data partition to query
          required: false
          schema:
            title: data partition id
            minLength: 1
            type: string
            description: identifier of the data partition to query
          name: data-partition-id
          in: header
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dipset'
        '404':
          description: DipSet not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - OpenDESBearerToken: []
  '/ddms/v2/dipsets/{dipsetid}/versions':
    get:
      tags:
        - Dipset
      summary: Get all versions of the dipset
      operationId: get_dipset_versions
      parameters:
        - required: true
          schema:
            title: Dipsetid
            type: string
          name: dipsetid
          in: path
        - description: identifier of the data partition to query
          required: false
          schema:
            title: data partition id
            minLength: 1
            type: string
            description: identifier of the data partition to query
          name: data-partition-id
          in: header
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordVersions'
        '404':
          description: DipSet not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - OpenDESBearerToken: []
  '/ddms/v2/dipsets/{dipsetid}':
    get:
      tags:
        - Dipset
      summary: 'Get the DipSet using wks:dipSet:1.0.0 schema'
      description: Get the DipSet object using its **id**
      operationId: get_dipset
      parameters:
        - required: true
          schema:
            title: Dipsetid
            type: string
          name: dipsetid
          in: path
        - description: identifier of the data partition to query
          required: false
          schema:
            title: data partition id
            minLength: 1
            type: string
            description: identifier of the data partition to query
          name: data-partition-id
          in: header
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/dipset'
        '404':
          description: DipSet not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - OpenDESBearerToken: []
    delete:
      tags:
        - Dipset
      summary: Delete the DipSet. The API performs a logical deletion of the given record
      operationId: del_dipset
      parameters:
        - required: true
          schema:
            title: Dipsetid
            type: string
          name: dipsetid
          in: path
        - description: Whether or not to delete records children
          required: false
          schema:
            title: Recursive
            type: boolean
            description: Whether or not to delete records children
            default: false
          name: recursive
          in: query
        - description: identifier of the data partition to query
          required: false
          schema:
            title: data partition id
            minLength: 1
            type: string
            description: identifier of the data partition to query
          name: data-partition-id
          in: header
      responses:
        '204':
          description: Record deleted successfully
        '404':
          description: DipSet not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - OpenDESBearerToken: []
...
components:
  schemas:
...
    CreateUpdateRecordsResponse:
      title: CreateUpdateRecordsResponse
      type: object
      properties:
        recordCount:
          title: Recordcount
          type: integer
        recordIds:
          title: Recordids
          type: array
          items:
            type: string
        skippedRecordIds:
          title: Skippedrecordids
          type: array
          items:
            type: string
    DataType:
      title: DataType
      enum:
        - string
        - number
        - integer
        - boolean
      description: An enumeration.
...
    Format:
      title: Format
      enum:
        - date
        - date-time
        - time
        - byte
        - binary
        - boolean
        - email
        - uuid
        - uri
        - int8
        - int16
        - int32
        - int64
        - float32
        - float64
        - float128
      description: An enumeration.
...
    HTTPValidationError:
      title: HTTPValidationError
      type: object
      properties:
        errors:
          title: Errors
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
    Kind:
      title: Kind
      enum:
        - CRS
        - Unit
        - Measurement
        - AzimuthReference
        - DateTime
      type: string
      description: An enumeration.
    LinkList:
      title: LinkList
      type: object
      properties: {}
      additionalProperties: false
      description: The base model forbids fields which are not declared initially in the pydantic model
...
    MetaItem:
      title: MetaItem
      required:
        - kind
        - persistableReference
      type: object
      properties:
        kind:
          title: Reference Kind
          allOf:
            - $ref: '#/components/schemas/Kind'
          description: 'The kind of reference, unit, measurement, CRS or azimuth reference.'
        name:
          title: Name or Symbol
          type: string
          description: The name of the CRS or the symbol/name of the unit
        persistableReference:
          title: Persistable Reference
          type: string
          description: The persistable reference string uniquely identifying the CRS or Unit
        propertyNames:
          title: Attribute Names
          type: array
          items:
            type: string
          description: 'The list of property names, to which this meta data item provides Unit/CRS context to. Data structures, which come in a single frame of reference, can register the property name, others require a full path like "data.structureA.propertyB" to define a unique context.'
        propertyValues:
          title: Attribute Names
          type: array
          items:
            type: string
          description: 'The list of property values, to which this meta data item provides Unit/CRS context to. Typically a unit symbol is a value to a data structure; this symbol is then registered in this propertyValues array and the persistableReference provides the absolute reference.'
        uncertainty:
          title: Uncertainty
          type: number
          description: The uncertainty of the values measured given the unit or CRS unit.
      additionalProperties: false
      description: The base model forbids fields which are not declared initially in the pydantic model
...
    RecordVersions:
      title: RecordVersions
      type: object
      properties:
        recordId:
          title: Recordid
          type: string
        versions:
          title: Versions
          type: array
          items:
            type: integer
...
    SimpleElevationReference:
      title: SimpleElevationReference
      required:
        - elevationFromMsl
      type: object
      properties:
        elevationFromMsl:
          title: Elevation from MSL
          allOf:
            - $ref: '#/components/schemas/ValueWithUnit'
          description: 'The elevation above mean sea level (MSL), at which the vertical origin is 0.0. The ''unitKey'' is further defined in ''frameOfReference.units''.'
        name:
          title: Elevation Reference Name
          type: string
          description: The name of the Elevation Reference.
      additionalProperties: false
      description: The base model forbids fields which are not declared initially in the pydantic model
...
    TagDictionary:
      title: TagDictionary
      type: object
      properties: {}
      description: Used for data model allows extra fields which are not declared initially in the pydantic model
    ToOneRelationship:
      title: ToOneRelationship
      type: object
      properties:
        confidence:
          title: Relationship Confidence
          type: number
          description: The confidence of the relationship. If the property is absent a well-known relation is implied.
        id:
          title: Related Object Id
          type: string
          description: 'The id of the related object in the Data Ecosystem. If set, the id has priority over the natural key in the name property.'
        name:
          title: Related Object Name
          type: string
          description: The name or natural key of the related object. This property is required if the target object id could not (yet) be identified.
        version:
          title: Entity Version Number
          type: number
          description: 'The version number of the related entity. If no version number is specified, the last version is implied.'
      additionalProperties: false
      description: The base model forbids fields which are not declared initially in the pydantic model
...
    ValidationError:
      title: ValidationError
      required:
        - loc
        - msg
        - type
      type: object
      properties:
        loc:
          title: Location
          type: array
          items:
            type: string
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
    ValueWithUnit:
      title: ValueWithUnit
      required:
        - unitKey
        - value
      type: object
      properties:
        unitKey:
          title: Unit Key
          type: string
          description: Unit for value of the corresponding attribute for the domain object in question. The key can be looked up in the 'frameOfReference.units' for further details.
        value:
          title: Value
          type: number
          description: Value of the corresponding attribute for the domain object in question.
      additionalProperties: false
      description: The base model forbids fields which are not declared initially in the pydantic model
...
    app__model__model_curated__Legal:
      title: Legal
      type: object
      properties:
        legaltags:
          title: Legal Tags
          type: array
          items:
            type: string
          description: 'The list of legal tags, see compliance API.'
        otherRelevantDataCountries:
          title: Other Relevant Data Countries
          type: array
          items:
            type: string
          description: 'The list of other relevant data countries using the ISO 2-letter codes, see compliance API.'
        status:
          title: Legal Status
          type: string
          description: The legal status.
      additionalProperties: false
      description: The base model forbids fields which are not declared initially in the pydantic model
...
    channel:
      title: channel
      type: object
      properties:
        absentValue:
          title: Absent Value
          type: string
          description: Optional field carrying the absent value as string for this channel.
        dataType:
          title: Data Type
          allOf:
            - $ref: '#/components/schemas/DataType'
          description: The log value type (per log sample). The 'format' property may contain further hints about data type presentation.
          default: number
        dimension:
          title: Dimension
          type: integer
          description: The dimension of this log or channel
        family:
          title: Log Family
          type: string
          description: The log family code of this log or channel (optional)
        familyType:
          title: Log Family Type
          type: string
          description: 'The log family type code of this log or channel. Example: ''Neutron Porosity'' for ''Thermal Neutron Porosity Sandstone''. (optional)'
        format:
          title: Format Hint
          allOf:
            - $ref: '#/components/schemas/Format'
          description: Optional format hint how to treat the log values as strings or number of bits per 'dataType'.
          default: float32
        logstoreId:
          title: Logstore ID
          type: number
          description: The id of this log or channel in the Logstore. This property is not present in the index channel.
        bulkURI:
          title: bulk URI
          type: string
          description: bulkURI either URL or URN.
        longName:
          title: Log Long Name
          type: string
          description: The long name of this log or channel
        mnemonic:
          title: Mnemonic
          type: string
          description: The mnemonic of this log or channel
        name:
          title: Log Name
          type: string
          description: The name of this log or channel.
        properties:
          title: Named Properties
          type: array
          items:
            $ref: '#/components/schemas/namedProperty'
          description: The named properties of this log or channel.
        source:
          title: Source
          type: string
          description: 'The source of this log or channel as a data reference; Typically this refers to the raw LogSet, from which this log WKE is generated.'
        unitKey:
          title: Unit
          type: string
          description: The unitKey to be looked up in the 'frameOfReference.units' dictionary to find the self-contained definition.
      additionalProperties: false
      description: The base model forbids fields which are not declared initially in the pydantic model
...
    dipSetData:
      title: dipSetData
      type: object
      properties:
        azimuthReference:
          title: Azimuth Reference Code
          type: string
          description: Azimuth reference code defining the type of North. Only used for dipSets with azimuth data
        classification:
          title: Log Set Classification
          type: string
          description: The well-known log set classification code.
          default: Externally Processed LogSet
        dateCreated:
          title: Creation Date and Time
          type: string
          description: The UTC date time of the entity creation
          format: date-time
        dateModified:
          title: Last Modification Date and Time
          type: string
          description: The UTC date time of the last entity modification
          format: date-time
        elevationReference:
          $ref: '#/components/schemas/SimpleElevationReference'
        externalIds:
          title: Array of External IDs
          type: array
          items:
            type: string
          description: 'An array of identities (e.g. some kind if URL to be resolved in an external data store), which links to external realizations of the same entity.'
        name:
          title: Dip Set Name
          type: string
          description: The name of this dip set
        operation:
          title: Operation
          type: string
          description: The operation which created this entity
        reference:
          $ref: '#/components/schemas/channel'
        referenceType:
          title: Reference Type
          type: string
          description: The reference index type of the dip set.
        relationships:
          $ref: '#/components/schemas/dipsetrelationships'
        start:
          $ref: '#/components/schemas/ValueWithUnit'
        step:
          $ref: '#/components/schemas/ValueWithUnit'
        stop:
          $ref: '#/components/schemas/ValueWithUnit'
        bulkURI:
          title: bulk URI
          type: string
          description: bulkURI either URL or URN.
      description: Used for data model allows extra fields which are not declared initially in the pydantic model
    dipset:
      title: dipset
      type: object
      properties:
        acl:
          title: Access Control List
          allOf:
            - $ref: '#/components/schemas/TagDictionary'
          description: The access control tags associated with this entity.
        ancestry:
          title: Ancestry
          allOf:
            - $ref: '#/components/schemas/LinkList'
          description: 'The links to data, which constitute the inputs.'
        data:
          title: Dip Set Data
          allOf:
            - $ref: '#/components/schemas/dipSetData'
          description: dipset data
        id:
          title: Dip Set ID
          type: string
          description: The unique identifier of the dip set
        kind:
          title: Dip Set Kind
          type: string
          description: Kind specification
          default: 'osdu:wks:dipSet:0.0.1'
        legal:
          title: Legal Tags
          allOf:
            - $ref: '#/components/schemas/app__model__model_curated__Legal'
          description: The dip-set's legal tags
        meta:
          title: Frame of Reference Meta Data
          type: array
          items:
            $ref: '#/components/schemas/MetaItem'
          description: 'The meta data section linking the ''unitKey'', ''crsKey'' to self-contained definitions (persistableReference)'
        type:
          title: Entity Type
          type: string
          description: 'The reference entity type as declared in common:metadata:entity:*.'
        version:
          title: Entity Version Number
          type: number
          description: The version number of this dip set; set by the framework.
      additionalProperties: false
      description: The base model forbids fields which are not declared initially in the pydantic model
    dipsetrelationships:
      title: dipsetrelationships
      required:
        - wellbore
      type: object
      properties:
        well:
          title: Well
          allOf:
            - $ref: '#/components/schemas/ToOneRelationship'
          description: The well to which this dipSet belongs. Only required if the wellbore is unknown.
        wellbore:
          title: Wellbore
          allOf:
            - $ref: '#/components/schemas/ToOneRelationship'
          description: The wellbore to which this dipSet belongs.
        wellboreSection:
          title: Wellbore Section
          allOf:
            - $ref: '#/components/schemas/ToOneRelationship'
          description: The wellbore section to which this dipSet belongs.
        referenceLog:
          title: True dip azimuth log
          allOf:
            - $ref: '#/components/schemas/ToOneRelationship'
          description: The true dip azimuth log of the dipset.
        trueDipAzimuthLog:
          title: True dip azimuth log
          allOf:
            - $ref: '#/components/schemas/ToOneRelationship'
          description: The true dip azimuth log of the dipset.
        trueDipInclinationLog:
          title: X-coordinate log
          allOf:
            - $ref: '#/components/schemas/ToOneRelationship'
          description: The X-coordinate log of the dipset
        xCoordinateLog:
          title: X-coordinate log
          allOf:
            - $ref: '#/components/schemas/ToOneRelationship'
          description: The X-coordinate log of the dipset
        yCoordinateLog:
          title: Y-coordinate log
          allOf:
            - $ref: '#/components/schemas/ToOneRelationship'
          description: The Y-coordinate log of the dipset
        zCoordinateLog:
          title: Z-coordinate log
          allOf:
            - $ref: '#/components/schemas/ToOneRelationship'
          description: The Z-coordinate log of the dipset
        qualityLog:
          title: Quality log
          allOf:
            - $ref: '#/components/schemas/ToOneRelationship'
          description: The quality log of the dipset
        classificationLog:
          title: Classification log
          allOf:
            - $ref: '#/components/schemas/ToOneRelationship'
          description: The classification log of the dipset
      additionalProperties: false
      description: The base model forbids fields which are not declared initially in the pydantic model
...
    namedProperty:
      title: namedProperty
      type: object
      properties:
        associations:
          title: Associations
          type: array
          items:
            type: string
          description: The optional associations contains one or more mnemonics found elsewhere in the logSet.
        description:
          title: Property Description
          type: string
          description: The description and role of this property.
        format:
          title: Format (LAS)
          type: string
          description: 'An optional format declaration for the property values. The ''A'' prefix indicates an array; string values are represented by ''S''; floating point values are represented by ''F'', optionally followed by a field specification, e.g. ''F10.4''; exponential number representations are represented by ''E''; integer values are represented by ''I''. For further information see the LAS specification http://www.cwls.org/las/.'
        name:
          title: Property Name
          type: string
          description: The name of this property.
        unitKey:
          title: Property Unit Symbol
          type: string
          description: The unitKey to be looked up in the 'frameOfReference.units' dictionary to find the self-contained definition.
        value:
          title: Property Value
          anyOf:
            - type: number
            - type: string
          description: The value for this property as a string or a number.
        values:
          title: Property Values (Interval)
          type: array
          items:
            type: number
          description: 'The values, e.g. interval boundaries, for this property.'
      additionalProperties: false
      description: The base model forbids fields which are not declared initially in the pydantic model
...
  securitySchemes:
    OpenDESBearerToken:
      type: http
      scheme: bearer